% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addComponent.R
\name{addComponent}
\alias{addComponent}
\title{Add new Component}
\usage{
addComponent(db_connection, componentTibble)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}

\item{componentTibble}{a `tibble` of the component information to upload}
}
\description{
Add a new component to the Access Level table
}
