% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioBankHash.R
\name{bioBankHash}
\alias{bioBankHash}
\title{BioBank Hash Code}
\usage{
bioBankHash(inFile)
}
\arguments{
\item{inFile}{a `tibble` from the file input of BioBank batch upload}
}
\value{
a character vector of the base64 encoded BioBank hash codes
}
\description{
A simple generation of a hash code which is used to check if a new BioBank record already has an existing entry in the database.
}
\details{
The hash code is simply the base64 encoded concatantion of; `Sample ID`, `BioFluid` and `SampleContainer`
}
