% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSampleTrackerHash.R
\name{checkSampleTrackerHash}
\alias{checkSampleTrackerHash}
\title{Check Sample Tracker Hash}
\usage{
checkSampleTrackerHash(formData, db_connection)
}
\arguments{
\item{formData}{a character of the reactive output from the Sample Tracker form entry}

\item{db_connection}{a `dbPool` or `RMariaDB` database connection}
}
\value{
a numeric value either 0 (hash already exists) 1 (incoming hash is new)
}
\description{
Check if the incoming Sample Tracker entry already exists in the database
}
