% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactMandatory.R
\name{contactMandatory}
\alias{contactMandatory}
\title{Mandatory Value Check (Contact Form)}
\usage{
contactMandatory(inputData)
}
\arguments{
\item{inputData}{list of input data from Participant Interest contact form}
}
\value{
a numeric value either 0;missing mandatory values or  1; all values present
}
\description{
Mandatory Value Check (Contact Form)
}
