% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBSform.R
\name{DBSform}
\alias{DBSform}
\title{DBS Form Data}
\usage{
DBSform(formData)
}
\arguments{
\item{formData}{a character of the reactive output from the Sample Tracker form entry}
}
\value{
return a character of `formData` for DBS cards
}
\description{
Extract Sample Tracker data for Dried Blood Spot (DBS) cards
}
