% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteSampleTracker.R
\name{deleteSampleTracker}
\alias{deleteSampleTracker}
\title{Delete Sample Tracker}
\usage{
deleteSampleTracker(db_connection, hash)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}

\item{hash}{a character stirng of the `hash` for the entry to delete}
}
\description{
Delete Sample Tracker
}
