% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegistrationHashCodes.R
\name{getRegistrationHashCodes}
\alias{getRegistrationHashCodes}
\title{Get Registration Hash Codes}
\usage{
getRegistrationHashCodes(db_connection)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}
}
\value{
a character vector of hash codes
}
\description{
Get Registration Hash Codes
}
