% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSourceComponents.R
\name{getSourceComponents}
\alias{getSourceComponents}
\title{Get Source Components}
\usage{
getSourceComponents(db_connection, username)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}

\item{username}{a character string of a valid `username`}
}
\value{
a vector of components to source at login
}
\description{
Retrieve the components to source at login based on the User's access level
}
