% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTabPanelNames.R
\name{getTabPanelNames}
\alias{getTabPanelNames}
\title{Get tabPanel Names}
\usage{
getTabPanelNames(db_connection, username)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}

\item{username}{a character string of a valid `username`}
}
\value{
a vector of components to source at login
}
\description{
Retrieve the names of `tabPanel` and `uiOutputs` for each component based the User's access level
}
