% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBioBank.R
\name{loadBioBank}
\alias{loadBioBank}
\title{Load BioBank}
\usage{
loadBioBank(db_connection)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}
}
\value{
a `tibble` of the complete BioBank table
}
\description{
Load the BioBak table
}
