% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBioBankIndex.R
\name{loadBioBankIndex}
\alias{loadBioBankIndex}
\title{Load BioBank Index}
\usage{
loadBioBankIndex(db_connection)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}
}
\value{
a `tibble` of BioBank box descriptions
}
\description{
Load the BioBank Index table
}
