% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadParticipantRegistration.R
\name{loadParticipantRegistration}
\alias{loadParticipantRegistration}
\title{Load Participant Registration Data}
\usage{
loadParticipantRegistration(db_connection)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}
}
\value{
a `tibble` of registered participants
}
\description{
Load Participant Registration Data
}
