% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProjectIndex.R
\name{loadProjectIndex}
\alias{loadProjectIndex}
\title{Load Project Index}
\usage{
loadProjectIndex(db_connection)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}
}
\value{
a `tibble` of WARU projets
}
\description{
Load the Project Index table
}
