% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRecruitmentContacts.R
\name{loadRecruitmentContacts}
\alias{loadRecruitmentContacts}
\title{Load Recruitment Contacts}
\usage{
loadRecruitmentContacts(db_connection)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}
}
\value{
a `tibble` of the `Participant Recruitment Database`
}
\description{
Load all data from the `Participant Recruitment Database`
}
