% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSampleTracker.R
\name{loadSampleTracker}
\alias{loadSampleTracker}
\title{Load Sample Tracker Data}
\usage{
loadSampleTracker(db_connection)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}
}
\value{
a `tibble` of sample details
}
\description{
Load useful data from *SampleTracker* table
}
