% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSampleTrackerAll.R
\name{loadSampleTrackerAll}
\alias{loadSampleTrackerAll}
\title{Load All Sample Tracker}
\usage{
loadSampleTrackerAll(db_connection)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}
}
\value{
a `tibble` of the entire `SampleTracker` table
}
\description{
Load All Sample Tracker
}
