% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSampleTrackerDelete.R
\name{loadSampleTrackerDelete}
\alias{loadSampleTrackerDelete}
\title{Load Sample Tracker (Deleted) Data}
\usage{
loadSampleTrackerDelete(db_connection)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}
}
\value{
a `tibble` of deleted sample details
}
\description{
Load useful data from *SampleTrackerDelete* table
}
