% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadUserData.R
\name{loadUserData}
\alias{loadUserData}
\title{Load User Data}
\usage{
loadUserData(db_connection)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}
}
\value{
a `tibble` of active users
}
\description{
Load all active users from *Users* table. Active refers to users who are currently registered for use on the system; not user who are currently active within a session.
}
