% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newProject.R
\name{newProject}
\alias{newProject}
\title{Add new Project}
\usage{
newProject(db_connection, projectTibble)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}

\item{projectTibble}{a `tibble` of the project information to upload}
}
\description{
Add a new project to the Project Index table
}
