% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBioBank.R
\name{prepareBioBank}
\alias{prepareBioBank}
\title{Prepare BioBank}
\usage{
prepareBioBank(inFile, username)
}
\arguments{
\item{inFile}{a `tibble` from the file input}

\item{username}{a character of the current session user (ie, `USER$name`)}
}
\value{
a `tibble` of the prepared `inFile`
}
\description{
Preparing incoming BioBank batch file for upload to the database
}
