% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareContactForm.R
\name{prepareContactForm}
\alias{prepareContactForm}
\title{Prepare Contact Infomation Form}
\usage{
prepareContactForm(formData, username)
}
\arguments{
\item{formData}{a character of the reactive output from the Contact Form entry}

\item{username}{a character of the current session user (ie, `USER$name`)}
}
\value{
a list of prepared form data
}
\description{
Prepare incoming Contact Form data for input into the database
}
