% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareContactFormBatch.R
\name{prepareContactFormBatch}
\alias{prepareContactFormBatch}
\title{Prepare Contact Infomation Form (Batch Mode)}
\usage{
prepareContactFormBatch(inFile, username)
}
\arguments{
\item{inFile}{a `tibble` from the file input}

\item{username}{a character of the current session user (ie, `USER$name`)}
}
\value{
a list of prepared form data
}
\description{
Prepare incoming Contact Form data for batch input into the database
}
