% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareRegistrationForm.R
\name{prepareRegistrationForm}
\alias{prepareRegistrationForm}
\title{Prepare New Participant Registration Form}
\usage{
prepareRegistrationForm(formData, username)
}
\arguments{
\item{formData}{a character of the reactive output from the Registration Form entry}

\item{username}{a character of the current session user (ie, `USER$name`)}
}
\value{
a list of prepared form data
}
\description{
Prepare incoming New Participant Registration Form data for input into the database
}
