% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectStatus.R
\name{projectStatus}
\alias{projectStatus}
\title{Change Project Status}
\usage{
projectStatus(db_connection, projectCode, projectStatus)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}

\item{projectCode}{a character string of the Project Code/Name}

\item{projectStatus}{a character string of either `Ongoing` or `Completed`}
}
\description{
Set the project status in the Proejct Index table to eiher **Ongoing** or **Completed**
}
