% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registrationHash.R
\name{registrationHash}
\alias{registrationHash}
\title{Registration Form Hash Code}
\usage{
registrationHash(formData)
}
\arguments{
\item{formData}{a character of the reactive output from the Registration Form entry}
}
\value{
a character string of the base64 encoded hash
}
\description{
A simple generation of a hash code which is used to check if newly registered participant has an existing entry in the database.
}
