% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registrationMandatory.R
\name{registrationMandatory}
\alias{registrationMandatory}
\title{Mandatory Value Check (Participant Registration Form)}
\usage{
registrationMandatory(inputData)
}
\arguments{
\item{inputData}{list of input data from Participant Registration form}
}
\value{
a numeric value either 0;missing mandatory values or  1; all values present
}
\description{
Mandatory Value Check (Participant Registration Form)
}
