% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleTrackerHash.R
\name{sampleTrackerHash}
\alias{sampleTrackerHash}
\title{Sample Tracker Hash Code}
\usage{
sampleTrackerHash(formData)
}
\arguments{
\item{formData}{a character of the reactive output from the Sample Tracker form entry}
}
\value{
a character of the base64 encoded sample hash code
}
\description{
A simple generation of a hash code which is used to check if an incoming sample already has an existing entry in the database.
}
\details{
The hash code is simply the base64 encoded concatantion of; `Sample ID`, `Date Received` and `Sample Type`
}
