% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleTrackerMandatory.R
\name{sampleTrackerMandatory}
\alias{sampleTrackerMandatory}
\title{Mandatory Value Check (Sample Tracker)}
\usage{
sampleTrackerMandatory(inputData)
}
\arguments{
\item{inputData}{list of input data from Sample Tracker form}
}
\value{
a numeric value either 0;missing mandatory values or  1; all values present
}
\description{
Mandatory Value Check (Sample Tracker)
}
