% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveBioBank.R
\name{saveBioBank}
\alias{saveBioBank}
\title{Save BioBank Data}
\usage{
saveBioBank(db_connection, inFile)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}

\item{inFile}{the `tibble` output from `prepareBioBank`}
}
\description{
Save BioBank records to the database
}
