% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSampleTrackerDelete.R
\name{saveSampleTrackerDelete}
\alias{saveSampleTrackerDelete}
\title{Save Sample Tracker (Delete)}
\usage{
saveSampleTrackerDelete(db_connection, inFile, username)
}
\arguments{
\item{db_connection}{a `dbPool` or `RMariaDB` database connection}

\item{inFile}{the `tibble` output from `loadSampleTracker`}

\item{username}{a character of the current session user (ie, `USER$name`)}
}
\description{
Save Sample Tracker (Delete)
}
