% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateBioBank.R
\name{validateBioBank}
\alias{validateBioBank}
\title{Validate Sample Tracker (Batch Upload)}
\usage{
validateBioBank(inFile, db_connection)
}
\arguments{
\item{inFile}{a `tibble` from the file input}

\item{db_connection}{a `dbPool` or `RMariaDB` database connection}
}
\value{
a numeric value either 0;validation failed or  1; validation successful
}
\description{
Validate the BioBank batch upload file
}
