% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateContactForm.R
\name{validateContactForm}
\alias{validateContactForm}
\title{Validate Contact Form (Batch Upload)}
\usage{
validateContactForm(inFile)
}
\arguments{
\item{inFile}{a `tibble` from the file input}
}
\value{
a numeric value either 0;validation failed or  1; validation successful
}
\description{
Validator function for when batch upload is used instead of standard form entry
}
