% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{validateEmail}
\alias{validateEmail}
\title{Validate E-Mail Address}
\usage{
validateEmail(x)
}
\arguments{
\item{x}{a character string of an E-Mail address}
}
\value{
a logical statement
}
\description{
Validate E-Mail Address
}
\keyword{internal}
