% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateSampleTracker.R
\name{validateSampleTracker}
\alias{validateSampleTracker}
\title{Validate Sample Tracker (Batch Upload)}
\usage{
validateSampleTracker(inFile, db_connection)
}
\arguments{
\item{inFile}{a `tibble` from the file input}

\item{db_connection}{a `dbPool` or `RMariaDB` database connection
#' @return a numeric value either 0;validation failed or  1; validation successful}
}
\description{
Validator function for when batch upload is used instead of standard form entry
}
