% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSampleTracker.R
\name{prepareSampleTracker}
\alias{prepareSampleTracker}
\title{Prepare Sample Tracker Form}
\usage{
prepareSampleTracker(formData, username)
}
\arguments{
\item{formData}{a character of the reactive output from the Sample Tracker form entry}

\item{username}{a character of the current session user (ie, `USER$name`)}
}
\value{
a character for prepared `formData` or a list of urine formData and DBS formData
}
\description{
Prepare incoming Sample Tracker form data for input into the database
}
