% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRegistrationForm.R
\name{saveRegistrationForm}
\alias{saveRegistrationForm}
\title{Register New Participant}
\usage{
saveRegistrationForm(formData, db_connection)
}
\arguments{
\item{formData}{the `list` output from `prepareRegistrationForm`}

\item{db_connection}{a `dbPool` or `RMariaDB` database connection}
}
\value{
a character string of the allocated ParticipantID
}
\description{
Register New Participant
}
